real-main/pyproject.toml
[tool.black]
line-length = 117
target-version = ['py38']
skip-string-normalization = true
exclude = '''
(
  /(
      \.git          # root of the project
    | node_modules
  )/
)
'''

[tool.poetry]
name = "real-main"
version = "0.1.0"
description = ""
authors = ["Mike Fogel <mike@real.app>"]

[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.dev-dependencies]
boto3 = "^1.11.9"
pytest = "^5.3.5"
requests-mock = "^1.7.0"
cryptography = "^2.8"
requests-aws4auth = "^0.9"
pillow = "^7.0.0"
elasticsearch = "^7.5.1"
more-itertools = "^8.2.0"
pendulum = "^2.0.5"
pytest-cov = "^2.8.1"
colorthief = "^0.2.1"
python-dotenv = "^0.12.0"
gql = "^0.4.0"
google-auth = "^1.12.0"
CacheControl = "^0.12.6"
pyheif = "^0.4"
pytest-xdist = "^1.31.0"
pytest-env = "^0.6.2"
aws-xray-sdk = "^2.5.0"
pytest-lazy-fixture = "^0.6.3"
moto = "1.3.15.dev859"
stringcase = "^1.2.0"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
